Add template transformation tests
